# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type MediaEntryKind = typeof MediaEntryKind[keyof typeof MediaEntryKind];


export const MediaEntryKind = {
  directory: 'directory',
  image: 'image',
  video: 'video',
  other: 'other',
  symlink: 'symlink',
} as const;

export interface MediaEntry {
  path: string;
  name: string;
  kind: MediaEntryKind;
  /** @minimum 0 */
  size: number;
  modifiedAt: string;
  /** @nullable */
  mediaType: string | null;
  tags: string[];
  /** @nullable */
  contentUrl: string | null;
}

export interface MediaDirectory {
  currentPath: string;
  /** @nullable */
  parentPath: string | null;
  entries: MediaEntry[];
  /** @minimum 0 */
  totalEntries: number;
  /** @minimum 0 */
  totalMedia: number;
}

export type TagOperationInputOperation = typeof TagOperationInputOperation[keyof typeof TagOperationInputOperation];


export const TagOperationInputOperation = {
  add: 'add',
  remove: 'remove',
  edit: 'edit',
} as const;

export interface TagOperationInput {
  operation: TagOperationInputOperation;
  /**
     * @minItems 1
     * @maxItems 250
     */
  paths: string[];
  /** @maxItems 50 */
  tags?: string[];
  oldTag?: string;
  newTag?: string;
}

export type TagFileResultStatus = typeof TagFileResultStatus[keyof typeof TagFileResultStatus];


export const TagFileResultStatus = {
  renamed: 'renamed',
  unchanged: 'unchanged',
} as const;

export interface TagFileResult {
  sourcePath: string;
  destinationPath: string;
  status: TagFileResultStatus;
}

export interface TagOperationResult {
  results: TagFileResult[];
  /** @minimum 0 */
  changedCount: number;
  /** @minimum 0 */
  unchangedCount: number;
}

export type FileTransferInputOperation = typeof FileTransferInputOperation[keyof typeof FileTransferInputOperation];


export const FileTransferInputOperation = {
  copy: 'copy',
  move: 'move',
} as const;

export interface FileTransferInput {
  operation: FileTransferInputOperation;
  /**
     * @minItems 1
     * @maxItems 250
     */
  paths: string[];
  destinationDirectory: string;
}

export type TransferFileResultStatus = typeof TransferFileResultStatus[keyof typeof TransferFileResultStatus];


export const TransferFileResultStatus = {
  copied: 'copied',
  moved: 'moved',
  failed: 'failed',
} as const;

export interface TransferFileResult {
  sourcePath: string;
  destinationPath: string;
  status: TransferFileResultStatus;
  /** @nullable */
  error: string | null;
}

export interface FileTransferResult {
  results: TransferFileResult[];
  /** @minimum 0 */
  successCount: number;
  /** @minimum 0 */
  failureCount: number;
}

export type GetMediaEntriesParams = {
/**
 * Directory path relative to the configured media root
 */
path?: string;
};

export type GetMediaContentParams = {
/**
 * File path relative to the configured media root
 */
path: string;
};