# lib/api-zod/src/generated/types/fileTransferInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FileTransferInputOperation } from './fileTransferInputOperation';

export interface FileTransferInput {
  operation: FileTransferInputOperation;
  /**
     * @minItems 1
     * @maxItems 250
     */
  paths: string[];
  destinationDirectory: string;
}