# lib/api-zod/src/generated/types/fileTransferResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransferFileResult } from './transferFileResult';

export interface FileTransferResult {
  results: TransferFileResult[];
  /** @minimum 0 */
  successCount: number;
  /** @minimum 0 */
  failureCount: number;
}