# lib/api-zod/src/generated/types/mediaDirectory.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MediaEntry } from './mediaEntry';

export interface MediaDirectory {
  currentPath: string;
  /** @nullable */
  parentPath: string | null;
  entries: MediaEntry[];
  /** @minimum 0 */
  totalEntries: number;
  /** @minimum 0 */
  totalMedia: number;
}