# lib/api-zod/src/generated/types/mediaEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MediaEntryKind } from './mediaEntryKind';

export interface MediaEntry {
  path: string;
  name: string;
  kind: MediaEntryKind;
  /** @minimum 0 */
  size: bigint;
  modifiedAt: Date;
  /** @nullable */
  mediaType: string | null;
  tags: string[];
  /** @nullable */
  contentUrl: string | null;
}