# lib/api-zod/src/generated/types/tagFileResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TagFileResultStatus } from './tagFileResultStatus';

export interface TagFileResult {
  sourcePath: string;
  destinationPath: string;
  status: TagFileResultStatus;
}