# lib/api-zod/src/generated/types/tagOperationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TagOperationInputOperation } from './tagOperationInputOperation';

export interface TagOperationInput {
  operation: TagOperationInputOperation;
  /**
     * @minItems 1
     * @maxItems 250
     */
  paths: string[];
  /** @maxItems 50 */
  tags?: string[];
  oldTag?: string;
  newTag?: string;
}