# lib/api-zod/src/generated/types/tagOperationResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TagFileResult } from './tagFileResult';

export interface TagOperationResult {
  results: TagFileResult[];
  /** @minimum 0 */
  changedCount: number;
  /** @minimum 0 */
  unchangedCount: number;
}