# lib/api-zod/src/generated/types/transferFileResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransferFileResultStatus } from './transferFileResultStatus';

export interface TransferFileResult {
  sourcePath: string;
  destinationPath: string;
  status: TransferFileResultStatus;
  /** @nullable */
  error: string | null;
}